ComputerAdaptiveTesting

Documentation for ComputerAdaptiveTesting.

ComputerAdaptiveTesting.Aggregators

ComputerAdaptiveTesting.IntegralCoeffs

ComputerAdaptiveTesting.Integrators

ComputerAdaptiveTesting.ItemBanks

ComputerAdaptiveTesting.NextItemRules

ComputerAdaptiveTesting.NextItemRules.catr_next_item_aliases

ComputerAdaptiveTesting.Aggregators.EnumerationOptimizer

ComputerAdaptiveTesting.Aggregators.RiemannEnumerationIntegrator

ComputerAdaptiveTesting.CatConfig.CatLoopConfig

ComputerAdaptiveTesting.CatConfig.CatRules

ComputerAdaptiveTesting.ItemBanks.CdfMirtItemBank

ComputerAdaptiveTesting.NextItemRules.AbilityVarianceStateCriterion

ComputerAdaptiveTesting.NextItemRules.DistributionExpectationBasedItemCriterion

ComputerAdaptiveTesting.NextItemRules.PointExpectationBasedItemCriterion

ComputerAdaptiveTesting.Aggregators.maybe_tracked_ability_estimate

ComputerAdaptiveTesting.ItemBanks.ItemBank2PL

ComputerAdaptiveTesting.ItemBanks.ItemBank3PL

ComputerAdaptiveTesting.ItemBanks.ItemBank4PL

ComputerAdaptiveTesting.ItemBanks.ItemBankMirt2PL

ComputerAdaptiveTesting.ItemBanks.ItemBankMirt3PL

ComputerAdaptiveTesting.ItemBanks.ItemBankMirt4PL

ComputerAdaptiveTesting.ItemBanks._search

ComputerAdaptiveTesting.ItemBanks.irf

ComputerAdaptiveTesting.Sim.auto_responder

ComputerAdaptiveTesting.Sim.prompt_response

ComputerAdaptiveTesting.Sim.run_cat

ComputerAdaptiveTesting.Aggregators — Module

This module takes care of integrating and optimizing over the ability/difficulty space. It includes TrackedResponses, which can store cumulative results during a test.

ComputerAdaptiveTesting.Aggregators.EnumerationOptimizer — Type

Argmax + max over the ability likihood given a set of responses with a given coefficient using exhaustive search.

TODO: Add item bank trait for enumerable item banks.

ComputerAdaptiveTesting.Aggregators.RiemannEnumerationIntegrator — Type

In case an item bank is enumerable (e.g. GriddedItemBank), then this method integrates over the ability likihood given a set of responses with a given coefficient function using a Riemann sum (aka the rectangle rule).

ComputerAdaptiveTesting.Aggregators.maybe_tracked_ability_estimate — Method

This method returns a tracked point estimate if it is has the given ability estimator, otherwise it computes it using the given ability estimator.

ComputerAdaptiveTesting.ItemBanks — Module

This module provides abstract and concrete item banks, which store information about items and their parameters such as difficulty, most typically resulting from fitting an Item-Response Theory (IRT) model. It includes TrackedResponses, which can store cumulative results during a test.

ComputerAdaptiveTesting.ItemBanks.CdfMirtItemBank — Type

This item bank corresponds to the most commonly found version of MIRT in the literature. Its items feature multidimensional discriminations and its learners multidimensional abilities, but item difficulties are single-dimensional.

ComputerAdaptiveTesting.ItemBanks.ItemBank2PL — Method

Convenience function to construct an item bank of the standard 2-parameter logistic single-dimensional IRT model.

ComputerAdaptiveTesting.ItemBanks.ItemBank3PL — Method

Convenience function to construct an item bank of the standard 3-parameter logistic single-dimensional IRT model.

ComputerAdaptiveTesting.ItemBanks.ItemBank4PL — Method

Convenience function to construct an item bank of the standard 4-parameter logistic single-dimensional IRT model.

ComputerAdaptiveTesting.ItemBanks.ItemBankMirt2PL — Method

Convenience function to construct an item bank of the standard 2-parameter logistic MIRT model.

ComputerAdaptiveTesting.ItemBanks.ItemBankMirt3PL — Method

Convenience function to construct an item bank of the standard 3-parameter logistic MIRT model.

ComputerAdaptiveTesting.ItemBanks.ItemBankMirt4PL — Method

Convenience function to construct an item bank of the standard 4-parameter logistic MIRT model.

ComputerAdaptiveTesting.ItemBanks._search — Method

Binary search for the point x where the integral from -inf...x is target += precis

ComputerAdaptiveTesting.ItemBanks.irf — Method

The item-response function at a particular item and value of θ. This is the empirical/nonparametric variant.

ComputerAdaptiveTesting.Sim.auto_responder — Method

This function constructs a next item function which automatically responds according to responses.

ComputerAdaptiveTesting.Sim.prompt_response — Method

This response callback simply prompts

ComputerAdaptiveTesting.Sim.run_cat — Method

Run a given CatLoopConfig

ComputerAdaptiveTesting.NextItemRules — Module

This module implements the next item selection rules, which form the main part of CAT.

Bibliography

[1] Linden, W. J., & Pashley, P. J. (2009). Item selection and ability estimation in adaptive testing. In Elements of adaptive testing (pp. 3-30). Springer, New York, NY.

ComputerAdaptiveTesting.NextItemRules.catr_next_item_aliases — Constant

This mapping provides next item rules through the same names that they are available through in the catR R package. TODO compability with mirtcat

ComputerAdaptiveTesting.NextItemRules.AbilityVarianceStateCriterion — Type

This StateCriterion returns the variance of the ability estimate given a set of responses.

ComputerAdaptiveTesting.NextItemRules.DistributionExpectationBasedItemCriterion — Type

This ItemCriterion wraps a StateCriterion and looks at its expected value for a particular item 1-ply ahead by integrating over an ability curve.

ComputerAdaptiveTesting.NextItemRules.PointExpectationBasedItemCriterion — Type

This ItemCriterion wraps a StateCriterion and looks at its expected value for a particular item 1-ply ahead based on a point ability estimate.

ComputerAdaptiveTesting.Integrators — Module

This module provides a common interface to different integration techniques.

ComputerAdaptiveTesting.IntegralCoeffs — Module

These are helpers for weighting integrals of p.d.f.s for calculating basic stats.

The main idea of doing it this way is to have a single instance of these to reuse specializations and to use structs so as to be able to control the level of specialization.

ComputerAdaptiveTesting.CatConfig.CatLoopConfig — Type

Configuration for a simulatable CAT.

ComputerAdaptiveTesting.CatConfig.CatRules — Type

Configuration of the rules for a CAT. This all includes all the basic rules for the CAT's operation, but not the item bank, nor any of the interactivity hooks needed to actually run the CAT.

This may be more a more convenient layer to integrate than CatLoopConfig if you want to write your own CAT loop rather than using hooks.

Modules = [CATPlots]