API reference

ComputerAdaptiveTesting.Aggregators

ComputerAdaptiveTesting.NextItemRules

ComputerAdaptiveTesting.NextItemRules.catr_next_item_aliases

ComputerAdaptiveTesting.Aggregators.RiemannEnumerationIntegrator

ComputerAdaptiveTesting.CatConfig.CatLoopConfig

ComputerAdaptiveTesting.CatConfig.CatRules

ComputerAdaptiveTesting.NextItemRules.AbilityVarianceStateCriterion

ComputerAdaptiveTesting.NextItemRules.DistributionExpectationBasedItemCriterion

ComputerAdaptiveTesting.NextItemRules.ExhaustiveSearch1Ply

ComputerAdaptiveTesting.NextItemRules.ItemCriterion

ComputerAdaptiveTesting.NextItemRules.ItemStrategyNextItemRule

ComputerAdaptiveTesting.NextItemRules.KernelAbstractionsExhaustiveSearch

ComputerAdaptiveTesting.NextItemRules.KernelAbstractionsExhaustiveSearchConfig

ComputerAdaptiveTesting.NextItemRules.NextItemRule

ComputerAdaptiveTesting.NextItemRules.NextItemStrategy

ComputerAdaptiveTesting.NextItemRules.PointExpectationBasedItemCriterion

ComputerAdaptiveTesting.NextItemRules.RandomNextItemRule

ComputerAdaptiveTesting.NextItemRules.StateCriterion

ComputerAdaptiveTesting.NextItemRules.UrryItemCriterion

ComputerAdaptiveTesting.TerminationConditions.FixedItemsTerminationCondition

ComputerAdaptiveTesting.TerminationConditions.TerminationCondition

ComputerAdaptiveTesting.Aggregators.maybe_tracked_ability_estimate

ComputerAdaptiveTesting.Sim.auto_responder

ComputerAdaptiveTesting.Sim.prompt_response

ComputerAdaptiveTesting.Sim.run_cat

ComputerAdaptiveTesting.Aggregators — Module

This module takes care of integrating and optimizing over the ability/difficulty space. It includes TrackedResponses, which can store cumulative results during a test.

ComputerAdaptiveTesting.Aggregators.RiemannEnumerationIntegrator — Type

In case an item bank is enumerable (e.g. GriddedItemBank), then this method integrates over the ability likihood given a set of responses with a given coefficient function using a Riemann sum (aka the rectangle rule).

ComputerAdaptiveTesting.Aggregators.maybe_tracked_ability_estimate — Method

This method returns a tracked point estimate if it is has the given ability estimator, otherwise it computes it using the given ability estimator.

ComputerAdaptiveTesting.Sim.auto_responder — Method

This function constructs a next item function which automatically responds according to responses.

ComputerAdaptiveTesting.Sim.prompt_response — Method

This response callback simply prompts

ComputerAdaptiveTesting.Sim.run_cat — Method

Run a given CatLoopConfig

ComputerAdaptiveTesting.TerminationConditions.FixedItemsTerminationCondition — Type

struct FixedItemsTerminationCondition <: ComputerAdaptiveTesting.TerminationConditions.TerminationCondition

num_items::Int64

ComputerAdaptiveTesting.TerminationConditions.TerminationCondition — Type

abstract type TerminationCondition <: ComputerAdaptiveTesting.ConfigBase.CatConfigBase

ComputerAdaptiveTesting.NextItemRules — Module

This module implements the next item selection rules, which form the main part of CAT.

Bibliography

[1] Linden, W. J., & Pashley, P. J. (2009). Item selection and ability estimation in adaptive testing. In Elements of adaptive testing (pp. 3-30). Springer, New York, NY.

ComputerAdaptiveTesting.NextItemRules.catr_next_item_aliases — Constant

This mapping provides next item rules through the same names that they are available through in the catR R package. TODO compability with mirtcat

ComputerAdaptiveTesting.NextItemRules.AbilityVarianceStateCriterion — Type

struct AbilityVarianceStateCriterion{DistEst<:ComputerAdaptiveTesting.Aggregators.DistributionAbilityEstimator, IntegratorT<:ComputerAdaptiveTesting.Aggregators.AbilityIntegrator} <: ComputerAdaptiveTesting.NextItemRules.StateCriterion

dist_est::ComputerAdaptiveTesting.Aggregators.DistributionAbilityEstimator

integrator::ComputerAdaptiveTesting.Aggregators.AbilityIntegrator

skip_zero::Bool

This StateCriterion returns the variance of the ability estimate given a set of responses.

ComputerAdaptiveTesting.NextItemRules.DistributionExpectationBasedItemCriterion — Type

struct DistributionExpectationBasedItemCriterion{DistributionAbilityEstimatorT<:ComputerAdaptiveTesting.Aggregators.DistributionAbilityEstimator, AbilityIntegratorT<:ComputerAdaptiveTesting.Aggregators.AbilityIntegrator, StateCriterionT<:ComputerAdaptiveTesting.NextItemRules.StateCriterion} <: ComputerAdaptiveTesting.NextItemRules.ExpectationBasedItemCriterion

ability_estimator::ComputerAdaptiveTesting.Aggregators.DistributionAbilityEstimator

integrator::ComputerAdaptiveTesting.Aggregators.AbilityIntegrator

state_criterion::ComputerAdaptiveTesting.NextItemRules.StateCriterion

This ItemCriterion wraps a StateCriterion and looks at its expected value for a particular item 1-ply ahead by integrating over an ability curve.

ComputerAdaptiveTesting.NextItemRules.ExhaustiveSearch1Ply — Type

struct ExhaustiveSearch1Ply <: ComputerAdaptiveTesting.NextItemRules.NextItemStrategy

parallel::Bool

ComputerAdaptiveTesting.NextItemRules.ItemCriterion — Type

abstract type ItemCriterion <: ComputerAdaptiveTesting.ConfigBase.CatConfigBase

ComputerAdaptiveTesting.NextItemRules.ItemStrategyNextItemRule — Type

struct ItemStrategyNextItemRule{NextItemStrategyT<:ComputerAdaptiveTesting.NextItemRules.NextItemStrategy, ItemCriterionT<:ComputerAdaptiveTesting.NextItemRules.ItemCriterion} <: ComputerAdaptiveTesting.NextItemRules.NextItemRule

strategy::ComputerAdaptiveTesting.NextItemRules.NextItemStrategy

criterion::ComputerAdaptiveTesting.NextItemRules.ItemCriterion

ItemStrategyNextItemRule which together with a NextItemStrategy acts as an adapter by which an ItemCriterion can serve as a NextItemRule.

ItemStrategyNextItemRule(bits...; ability_estimator=nothing, parallel=true)

Implicit constructor for ItemStrategyNextItemRule. Will default to ExhaustiveSearch1Ply when no NextItemStrategy is given.

ComputerAdaptiveTesting.NextItemRules.KernelAbstractionsExhaustiveSearch — Type

struct KernelAbstractionsExhaustiveSearch{KernelT} <: ComputerAdaptiveTesting.NextItemRules.NextItemStrategy

kernel::Any

ComputerAdaptiveTesting.NextItemRules.KernelAbstractionsExhaustiveSearchConfig — Type

struct KernelAbstractionsExhaustiveSearchConfig{ArgsT} <: ComputerAdaptiveTesting.NextItemRules.NextItemStrategy

kernel_args::Any

ComputerAdaptiveTesting.NextItemRules.NextItemRule — Type

abstract type NextItemRule <: ComputerAdaptiveTesting.ConfigBase.CatConfigBase

Abstract base type for all item selection rules. All descendants of this type are expected to implement the interface (rule::NextItemRule)(responses::TrackedResponses, items::AbstractItemBank)::Int

NextItemRule(bits...; ability_estimator=nothing, parallel=true)

Implicit constructor for NextItemRule. Uses any given NextItemRule or delegates to ItemStrategyNextItemRule.

ComputerAdaptiveTesting.NextItemRules.NextItemStrategy — Type

abstract type NextItemStrategy <: ComputerAdaptiveTesting.ConfigBase.CatConfigBase

ComputerAdaptiveTesting.NextItemRules.PointExpectationBasedItemCriterion — Type

struct PointExpectationBasedItemCriterion{PointAbilityEstimatorT<:ComputerAdaptiveTesting.Aggregators.PointAbilityEstimator, StateCriterionT<:ComputerAdaptiveTesting.NextItemRules.StateCriterion} <: ComputerAdaptiveTesting.NextItemRules.ExpectationBasedItemCriterion

ability_estimator::ComputerAdaptiveTesting.Aggregators.PointAbilityEstimator

state_criterion::ComputerAdaptiveTesting.NextItemRules.StateCriterion

This ItemCriterion wraps a StateCriterion and looks at its expected value for a particular item 1-ply ahead based on a point ability estimate.

ComputerAdaptiveTesting.NextItemRules.RandomNextItemRule — Type

struct RandomNextItemRule{RandomT<:Random.AbstractRNG} <: ComputerAdaptiveTesting.NextItemRules.NextItemRule

rng::Random.AbstractRNG: Default: Xoshiro()

This is the most basic rule for choosing the next item in a CAT. It simply picks a random item from the set of items that have not yet been administered.

ComputerAdaptiveTesting.NextItemRules.StateCriterion — Type

abstract type StateCriterion <: ComputerAdaptiveTesting.ConfigBase.CatConfigBase

ComputerAdaptiveTesting.NextItemRules.UrryItemCriterion — Type

struct UrryItemCriterion{AbilityEstimatorT<:ComputerAdaptiveTesting.Aggregators.PointAbilityEstimator} <: ComputerAdaptiveTesting.NextItemRules.ItemCriterion

ability_estimator::ComputerAdaptiveTesting.Aggregators.PointAbilityEstimator

This item criterion just picks the item with the raw difficulty closest to the current ability estimate.

ComputerAdaptiveTesting.CatConfig.CatLoopConfig — Type

Configuration for a simulatable CAT.

ComputerAdaptiveTesting.CatConfig.CatRules — Type

struct CatRules{NextItemRuleT<:ComputerAdaptiveTesting.NextItemRules.NextItemRule, TerminationConditionT<:ComputerAdaptiveTesting.TerminationConditions.TerminationCondition, AbilityEstimatorT<:ComputerAdaptiveTesting.Aggregators.AbilityEstimator, AbilityTrackerT<:ComputerAdaptiveTesting.Aggregators.AbilityTracker} <: ComputerAdaptiveTesting.ConfigBase.CatConfigBase

next_item::ComputerAdaptiveTesting.NextItemRules.NextItemRule: The rule to choose the next item in the CAT given the current state.

termination_condition::ComputerAdaptiveTesting.TerminationConditions.TerminationCondition: The rule to choose when to terminate the CAT.

ability_estimator::ComputerAdaptiveTesting.Aggregators.AbilityEstimator: The ability estimator, which estimates the testee's current ability.

ability_tracker::ComputerAdaptiveTesting.Aggregators.AbilityTracker: The ability tracker, which tracks the testee's current ability level. Default: NullAbilityTracker()

Configuration of the rules for a CAT. This all includes all the basic rules for the CAT's operation, but not the item bank, nor any of the interactivity hooks needed to actually run the CAT.

This may be more a more convenient layer to integrate than CatLoopConfig if you want to write your own CAT loop rather than using hooks.

CatRules(; next_item=..., termination_condition=..., ability_estimator=..., ability_tracker=...)

Explicit constructor for CatRules.

CatRules(bits...)

Implicit constructor for CatRules.